Calculate GAT Exclusions

This function cleans up the code to flag areas based on exclusion criteria so that they will not be used in the merge.

calculateGATexclusions(d, var, math, val)

Arguments

d
The dataset containing the flag and exclusion variables.

var
The variable to use when calculating exclusions.

math
The condition to use when calculating exclusions. Recognized conditions include "equals", "less than", "greater than".

val
The value to compare to var based on the condition selected.

Examples

calculateGATexclusions(
  d = hftown@data,
  var = "TOTAL_POP",
  math = "less than",
  val = 500
)

#>              TOWN         ID COUNTY   AREALAND  AREAWATR TOTAL_POP MY_FLAG
#> 1        Bleecker 035\\06882    035  148175370   5831674       573       0
#> 2      Broadalbin 035\\08433    035   82166307  20853757      5066       0
#> 3          Caroga 035\\12573    035  131696633   8774235      1407       1
#> 4        Ephratah 035\\24603    035  101576541    578556      1693       0
#> 5    Gloversville 035\\29443    035   13187984     13599     15413       0
#> 6  Johnstown City 035\\38781    035   12588379     16814      8511       0
#> 7       Johnstown 035\\38792    035  181852496   3009850      7166       0
#> 8        Mayfield 035\\46217    035  151343307  16212224      6432       0
#> 9     Northampton 035\\51407    035   54516993  35414661      2760       0
#> 10      Oppenheim 035\\55101    035  145745641    478206      1774       0
#> 11          Perth 035\\57353    035   67496451     54491      3638       0
#> 12      Stratford 035\\71784    035  194718598   3898019       640       0
#> 13        Arietta 041\\02572    041  822569831  30582930       293       0
#> 14         Benson 041\\05958    041  214097112   1314023       201       0
#> 15           Hope 041\\35496    041  105514427   2260160       392       0
#> 16    Indian Lake 041\\37374    041  654829452  34714986      1471       0
#> 17          Inlet 041\\37495    041  161328594  10636560       406       0
#> 18  Lake Pleasant 041\\40794    041  487129232  25623817       876       0
#> 19      Long Lake 041\\43412    041 1056388674 108850231       852       0
#> 20      Morehouse 041\\48340    041  494967545   9388815       151       0
#> 21          Wells 041\\79059    041  458973240   3030939       737       0
#>    GATflag
#> 1        0
#> 2        0
#> 3        0
#> 4        0
#> 5        0
#> 6        0
#> 7        0
#> 8        0
#> 9        0
#> 10       0
#> 11       0
#> 12       0
#> 13       1
#> 14       1
#> 15       1
#> 16       0
#> 17       1
#> 18       0
#> 19       0
#> 20       1
#> 21       0