Write GAT KML File

This function writes a KML file of the spatial layer.

writeGATkml(myshp, filename, filepath, myidvar = "GEOID10")

Arguments

myshp

A spatial layer.

filename

The desired name for the KML file.

filepath

The desired location for the KML file.

myidvar

A variable of unique values for the spatial layer.

Examples

# creates the kml file in the project root directory
if (interactive()) {
writeGATkml(myshp = hftown, filename = "my_kml_example",
            filepath = getwd(), myidvar = "TOWN")
}
#> Warning: number of items to replace is not a multiple of replacement length
#> Warning: number of items to replace is not a multiple of replacement length
#> Warning: number of items to replace is not a multiple of replacement length
#> Warning: number of items to replace is not a multiple of replacement length
#> Warning: number of items to replace is not a multiple of replacement length
#> Warning: number of items to replace is not a multiple of replacement length
#> Warning: number of items to replace is not a multiple of replacement length
#> Warning: number of items to replace is not a multiple of replacement length
#> Warning: number of items to replace is not a multiple of replacement length
#> Warning: number of items to replace is not a multiple of replacement length
#> Warning: number of items to replace is not a multiple of replacement length
#> Warning: number of items to replace is not a multiple of replacement length
#> Warning: number of items to replace is not a multiple of replacement length
#> Warning: number of items to replace is not a multiple of replacement length
#> Warning: number of items to replace is not a multiple of replacement length
#> Warning: number of items to replace is not a multiple of replacement length
#> Warning: number of items to replace is not a multiple of replacement length
#> Warning: number of items to replace is not a multiple of replacement length
#> Warning: number of items to replace is not a multiple of replacement length
#> Warning: number of items to replace is not a multiple of replacement length
#> Warning: number of items to replace is not a multiple of replacement length
#> Deleting source `C:/Users/ajstamm/Documents/GitHub/Coding/gatpkg/docs/dev/reference/my_kml_example.kml' using driver `KML'
#> Writing layer `my_kml_example' to data source 
#>   `C:/Users/ajstamm/Documents/GitHub/Coding/gatpkg/docs/dev/reference/my_kml_example.kml' using driver `KML'
#> Writing 21 features with 9 fields and geometry type Multi Polygon.
#> Warning: GDAL Message 1: Value 'Bleecker' of field my_kml_example.TOTAL_POP parsed incompletely to integer 0.
#> Warning: GDAL Message 1: Value '<table border = 1>  <tr> <td> TOWN </td> <td> Bleecker </td> </tr> <tr> <td> ID </td> <td> 035\06882 </td> </tr> <tr> <td> COUNTY </td> <td> 035 </td> </tr> <tr> <td> AREALAND </td> <td> 148175370 </td> </tr> <tr> <td> AREAWATR </td> <td> 5831674 </td> </tr> <tr> <td> TOTAL_POP </td> <td> 573 </td> </tr> <tr> <td> MY_FLAG </td> <td> 0 </td> </tr> </table>' of field my_kml_example.MY_FLAG parsed incompletely to real 0.
#> Warning: GDAL Message 1: Value 'Broadalbin' of field my_kml_example.TOTAL_POP parsed incompletely to integer 0.
#> Warning: GDAL Message 1: Value '<table border = 1>  <tr> <td> TOWN </td> <td> Broadalbin </td> </tr> <tr> <td> ID </td> <td> 035\08433 </td> </tr> <tr> <td> COUNTY </td> <td> 035 </td> </tr> <tr> <td> AREALAND </td> <td> 82166307 </td> </tr> <tr> <td> AREAWATR </td> <td> 20853757 </td> </tr> <tr> <td> TOTAL_POP </td> <td> 5066 </td> </tr> <tr> <td> MY_FLAG </td> <td> 0 </td> </tr> </table>' of field my_kml_example.MY_FLAG parsed incompletely to real 0.
#> Warning: GDAL Message 1: Value 'Caroga' of field my_kml_example.TOTAL_POP parsed incompletely to integer 0.
#> Warning: GDAL Message 1: Value '<table border = 1>  <tr> <td> TOWN </td> <td> Caroga </td> </tr> <tr> <td> ID </td> <td> 035\12573 </td> </tr> <tr> <td> COUNTY </td> <td> 035 </td> </tr> <tr> <td> AREALAND </td> <td> 131696633 </td> </tr> <tr> <td> AREAWATR </td> <td> 8774235 </td> </tr> <tr> <td> TOTAL_POP </td> <td> 1407 </td> </tr> <tr> <td> MY_FLAG </td> <td> 1 </td> </tr> </table>' of field my_kml_example.MY_FLAG parsed incompletely to real 0.
#> Warning: GDAL Message 1: Value 'Ephratah' of field my_kml_example.TOTAL_POP parsed incompletely to integer 0.
#> Warning: GDAL Message 1: Value '<table border = 1>  <tr> <td> TOWN </td> <td> Ephratah </td> </tr> <tr> <td> ID </td> <td> 035\24603 </td> </tr> <tr> <td> COUNTY </td> <td> 035 </td> </tr> <tr> <td> AREALAND </td> <td> 101576541 </td> </tr> <tr> <td> AREAWATR </td> <td> 578556 </td> </tr> <tr> <td> TOTAL_POP </td> <td> 1693 </td> </tr> <tr> <td> MY_FLAG </td> <td> 0 </td> </tr> </table>' of field my_kml_example.MY_FLAG parsed incompletely to real 0.
#> Warning: GDAL Message 1: Value 'Gloversville' of field my_kml_example.TOTAL_POP parsed incompletely to integer 0.
#> Warning: GDAL Message 1: Value '<table border = 1>  <tr> <td> TOWN </td> <td> Gloversville </td> </tr> <tr> <td> ID </td> <td> 035\29443 </td> </tr> <tr> <td> COUNTY </td> <td> 035 </td> </tr> <tr> <td> AREALAND </td> <td> 13187984 </td> </tr> <tr> <td> AREAWATR </td> <td> 13599 </td> </tr> <tr> <td> TOTAL_POP </td> <td> 15413 </td> </tr> <tr> <td> MY_FLAG </td> <td> 0 </td> </tr> </table>' of field my_kml_example.MY_FLAG parsed incompletely to real 0.
#> Warning: GDAL Message 1: Value 'Johnstown City' of field my_kml_example.TOTAL_POP parsed incompletely to integer 0.
#> Warning: GDAL Message 1: Value '<table border = 1>  <tr> <td> TOWN </td> <td> Johnstown City </td> </tr> <tr> <td> ID </td> <td> 035\38781 </td> </tr> <tr> <td> COUNTY </td> <td> 035 </td> </tr> <tr> <td> AREALAND </td> <td> 12588379 </td> </tr> <tr> <td> AREAWATR </td> <td> 16814 </td> </tr> <tr> <td> TOTAL_POP </td> <td> 8511 </td> </tr> <tr> <td> MY_FLAG </td> <td> 0 </td> </tr> </table>' of field my_kml_example.MY_FLAG parsed incompletely to real 0.
#> Warning: GDAL Message 1: Value 'Johnstown' of field my_kml_example.TOTAL_POP parsed incompletely to integer 0.
#> Warning: GDAL Message 1: Value '<table border = 1>  <tr> <td> TOWN </td> <td> Johnstown </td> </tr> <tr> <td> ID </td> <td> 035\38792 </td> </tr> <tr> <td> COUNTY </td> <td> 035 </td> </tr> <tr> <td> AREALAND </td> <td> 181852496 </td> </tr> <tr> <td> AREAWATR </td> <td> 3009850 </td> </tr> <tr> <td> TOTAL_POP </td> <td> 7166 </td> </tr> <tr> <td> MY_FLAG </td> <td> 0 </td> </tr> </table>' of field my_kml_example.MY_FLAG parsed incompletely to real 0.
#> Warning: GDAL Message 1: Value 'Mayfield' of field my_kml_example.TOTAL_POP parsed incompletely to integer 0.
#> Warning: GDAL Message 1: Value '<table border = 1>  <tr> <td> TOWN </td> <td> Mayfield </td> </tr> <tr> <td> ID </td> <td> 035\46217 </td> </tr> <tr> <td> COUNTY </td> <td> 035 </td> </tr> <tr> <td> AREALAND </td> <td> 151343307 </td> </tr> <tr> <td> AREAWATR </td> <td> 16212224 </td> </tr> <tr> <td> TOTAL_POP </td> <td> 6432 </td> </tr> <tr> <td> MY_FLAG </td> <td> 0 </td> </tr> </table>' of field my_kml_example.MY_FLAG parsed incompletely to real 0.
#> Warning: GDAL Message 1: Value 'Northampton' of field my_kml_example.TOTAL_POP parsed incompletely to integer 0.
#> Warning: GDAL Message 1: Value '<table border = 1>  <tr> <td> TOWN </td> <td> Northampton </td> </tr> <tr> <td> ID </td> <td> 035\51407 </td> </tr> <tr> <td> COUNTY </td> <td> 035 </td> </tr> <tr> <td> AREALAND </td> <td> 54516993 </td> </tr> <tr> <td> AREAWATR </td> <td> 35414661 </td> </tr> <tr> <td> TOTAL_POP </td> <td> 2760 </td> </tr> <tr> <td> MY_FLAG </td> <td> 0 </td> </tr> </table>' of field my_kml_example.MY_FLAG parsed incompletely to real 0.
#> Warning: GDAL Message 1: Value 'Oppenheim' of field my_kml_example.TOTAL_POP parsed incompletely to integer 0.
#> Warning: GDAL Message 1: Value '<table border = 1>  <tr> <td> TOWN </td> <td> Oppenheim </td> </tr> <tr> <td> ID </td> <td> 035\55101 </td> </tr> <tr> <td> COUNTY </td> <td> 035 </td> </tr> <tr> <td> AREALAND </td> <td> 145745641 </td> </tr> <tr> <td> AREAWATR </td> <td> 478206 </td> </tr> <tr> <td> TOTAL_POP </td> <td> 1774 </td> </tr> <tr> <td> MY_FLAG </td> <td> 0 </td> </tr> </table>' of field my_kml_example.MY_FLAG parsed incompletely to real 0.
#> Warning: GDAL Message 1: Value 'Perth' of field my_kml_example.TOTAL_POP parsed incompletely to integer 0.
#> Warning: GDAL Message 1: Value '<table border = 1>  <tr> <td> TOWN </td> <td> Perth </td> </tr> <tr> <td> ID </td> <td> 035\57353 </td> </tr> <tr> <td> COUNTY </td> <td> 035 </td> </tr> <tr> <td> AREALAND </td> <td> 67496451 </td> </tr> <tr> <td> AREAWATR </td> <td> 54491 </td> </tr> <tr> <td> TOTAL_POP </td> <td> 3638 </td> </tr> <tr> <td> MY_FLAG </td> <td> 0 </td> </tr> </table>' of field my_kml_example.MY_FLAG parsed incompletely to real 0.
#> Warning: GDAL Message 1: Value 'Stratford' of field my_kml_example.TOTAL_POP parsed incompletely to integer 0.
#> Warning: GDAL Message 1: Value '<table border = 1>  <tr> <td> TOWN </td> <td> Stratford </td> </tr> <tr> <td> ID </td> <td> 035\71784 </td> </tr> <tr> <td> COUNTY </td> <td> 035 </td> </tr> <tr> <td> AREALAND </td> <td> 194718598 </td> </tr> <tr> <td> AREAWATR </td> <td> 3898019 </td> </tr> <tr> <td> TOTAL_POP </td> <td> 640 </td> </tr> <tr> <td> MY_FLAG </td> <td> 0 </td> </tr> </table>' of field my_kml_example.MY_FLAG parsed incompletely to real 0.
#> Warning: GDAL Message 1: Value 'Arietta' of field my_kml_example.TOTAL_POP parsed incompletely to integer 0.
#> Warning: GDAL Message 1: Value '<table border = 1>  <tr> <td> TOWN </td> <td> Arietta </td> </tr> <tr> <td> ID </td> <td> 041\02572 </td> </tr> <tr> <td> COUNTY </td> <td> 041 </td> </tr> <tr> <td> AREALAND </td> <td> 822569831 </td> </tr> <tr> <td> AREAWATR </td> <td> 30582930 </td> </tr> <tr> <td> TOTAL_POP </td> <td> 293 </td> </tr> <tr> <td> MY_FLAG </td> <td> 0 </td> </tr> </table>' of field my_kml_example.MY_FLAG parsed incompletely to real 0.
#> Warning: GDAL Message 1: Value 'Benson' of field my_kml_example.TOTAL_POP parsed incompletely to integer 0.
#> Warning: GDAL Message 1: Value '<table border = 1>  <tr> <td> TOWN </td> <td> Benson </td> </tr> <tr> <td> ID </td> <td> 041\05958 </td> </tr> <tr> <td> COUNTY </td> <td> 041 </td> </tr> <tr> <td> AREALAND </td> <td> 214097112 </td> </tr> <tr> <td> AREAWATR </td> <td> 1314023 </td> </tr> <tr> <td> TOTAL_POP </td> <td> 201 </td> </tr> <tr> <td> MY_FLAG </td> <td> 0 </td> </tr> </table>' of field my_kml_example.MY_FLAG parsed incompletely to real 0.
#> Warning: GDAL Message 1: Value 'Hope' of field my_kml_example.TOTAL_POP parsed incompletely to integer 0.
#> Warning: GDAL Message 1: Value '<table border = 1>  <tr> <td> TOWN </td> <td> Hope </td> </tr> <tr> <td> ID </td> <td> 041\35496 </td> </tr> <tr> <td> COUNTY </td> <td> 041 </td> </tr> <tr> <td> AREALAND </td> <td> 105514427 </td> </tr> <tr> <td> AREAWATR </td> <td> 2260160 </td> </tr> <tr> <td> TOTAL_POP </td> <td> 392 </td> </tr> <tr> <td> MY_FLAG </td> <td> 0 </td> </tr> </table>' of field my_kml_example.MY_FLAG parsed incompletely to real 0.
#> Warning: GDAL Message 1: Value 'Indian Lake' of field my_kml_example.TOTAL_POP parsed incompletely to integer 0.
#> Warning: GDAL Message 1: Value '<table border = 1>  <tr> <td> TOWN </td> <td> Indian Lake </td> </tr> <tr> <td> ID </td> <td> 041\37374 </td> </tr> <tr> <td> COUNTY </td> <td> 041 </td> </tr> <tr> <td> AREALAND </td> <td> 654829452 </td> </tr> <tr> <td> AREAWATR </td> <td> 34714986 </td> </tr> <tr> <td> TOTAL_POP </td> <td> 1471 </td> </tr> <tr> <td> MY_FLAG </td> <td> 0 </td> </tr> </table>' of field my_kml_example.MY_FLAG parsed incompletely to real 0.
#> Warning: GDAL Message 1: Value 'Inlet' of field my_kml_example.TOTAL_POP parsed incompletely to integer 0.
#> Warning: GDAL Message 1: Value '<table border = 1>  <tr> <td> TOWN </td> <td> Inlet </td> </tr> <tr> <td> ID </td> <td> 041\37495 </td> </tr> <tr> <td> COUNTY </td> <td> 041 </td> </tr> <tr> <td> AREALAND </td> <td> 161328594 </td> </tr> <tr> <td> AREAWATR </td> <td> 10636560 </td> </tr> <tr> <td> TOTAL_POP </td> <td> 406 </td> </tr> <tr> <td> MY_FLAG </td> <td> 0 </td> </tr> </table>' of field my_kml_example.MY_FLAG parsed incompletely to real 0.
#> Warning: GDAL Message 1: Value 'Lake Pleasant' of field my_kml_example.TOTAL_POP parsed incompletely to integer 0.
#> Warning: GDAL Message 1: Value '<table border = 1>  <tr> <td> TOWN </td> <td> Lake Pleasant </td> </tr> <tr> <td> ID </td> <td> 041\40794 </td> </tr> <tr> <td> COUNTY </td> <td> 041 </td> </tr> <tr> <td> AREALAND </td> <td> 487129232 </td> </tr> <tr> <td> AREAWATR </td> <td> 25623817 </td> </tr> <tr> <td> TOTAL_POP </td> <td> 876 </td> </tr> <tr> <td> MY_FLAG </td> <td> 0 </td> </tr> </table>' of field my_kml_example.MY_FLAG parsed incompletely to real 0.
#> Warning: GDAL Message 1: Value 'Long Lake' of field my_kml_example.TOTAL_POP parsed incompletely to integer 0.
#> Warning: GDAL Message 1: Value '<table border = 1>  <tr> <td> TOWN </td> <td> Long Lake </td> </tr> <tr> <td> ID </td> <td> 041\43412 </td> </tr> <tr> <td> COUNTY </td> <td> 041 </td> </tr> <tr> <td> AREALAND </td> <td> 1056388674 </td> </tr> <tr> <td> AREAWATR </td> <td> 108850231 </td> </tr> <tr> <td> TOTAL_POP </td> <td> 852 </td> </tr> <tr> <td> MY_FLAG </td> <td> 0 </td> </tr> </table>' of field my_kml_example.MY_FLAG parsed incompletely to real 0.
#> Warning: GDAL Message 1: Value 'Morehouse' of field my_kml_example.TOTAL_POP parsed incompletely to integer 0.
#> Warning: GDAL Message 1: Value '<table border = 1>  <tr> <td> TOWN </td> <td> Morehouse </td> </tr> <tr> <td> ID </td> <td> 041\48340 </td> </tr> <tr> <td> COUNTY </td> <td> 041 </td> </tr> <tr> <td> AREALAND </td> <td> 494967545 </td> </tr> <tr> <td> AREAWATR </td> <td> 9388815 </td> </tr> <tr> <td> TOTAL_POP </td> <td> 151 </td> </tr> <tr> <td> MY_FLAG </td> <td> 0 </td> </tr> </table>' of field my_kml_example.MY_FLAG parsed incompletely to real 0.
#> Warning: GDAL Message 1: Value 'Wells' of field my_kml_example.TOTAL_POP parsed incompletely to integer 0.
#> Warning: GDAL Message 1: Value '<table border = 1>  <tr> <td> TOWN </td> <td> Wells </td> </tr> <tr> <td> ID </td> <td> 041\79059 </td> </tr> <tr> <td> COUNTY </td> <td> 041 </td> </tr> <tr> <td> AREALAND </td> <td> 458973240 </td> </tr> <tr> <td> AREAWATR </td> <td> 3030939 </td> </tr> <tr> <td> TOTAL_POP </td> <td> 737 </td> </tr> <tr> <td> MY_FLAG </td> <td> 0 </td> </tr> </table>' of field my_kml_example.MY_FLAG parsed incompletely to real 0.